Enhanced and moved dock menu to darwin folder and enabled mpris on macOS to support play/pause state
