+++ src/main/features/darwin/index.ts
@@ -1,0 +1,1 @@
+import './dock-menu';
